% Generated by roxygen2 (4.0.0): do not edit by hand
\name{answerlist}
\alias{answerlist}
\alias{format.answerlist}
\title{\code{answerlist} objects}
\usage{
answerlist(x)

\method{format}{answerlist}(x, show_solutions = FALSE,
  tpl_answerlist = examiner_opts$tpl_answerlist, format_i = identity,
  .debug = FALSE, ...)
}
\arguments{
\item{x}{A data frame with colums: text, correct, and fixed. In \code{format}, the object.}

\item{show_solutions}{Whether to the show the solution}

\item{tpl_answerlist}{The template to use to render the object.}

\item{format_i}{A function used to format the answer number.}

\item{.debug}{Show the data used in the template.}

\item{...}{Passed to the template.}
}
\value{
An object of class \code{answerlist}.
}
\description{
\code{answerlist} objects
}
\details{
The class \code{"answerlist"} is a \code{"data.frame"} with columns:
\tabular{lll}{
text \tab character \tab Answer text \cr
corect \tab logical \tab Is the answer correct? \cr
fixed \tab logical \tab Can the answer be shuffled
}
}

