% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_text_var.R
\name{clean_text_var}
\alias{clean_text_var}
\title{clean_text_var}
\usage{
clean_text_var(
  data = data,
  text_var = message,
  tolower = TRUE,
  remove_hashtags = TRUE,
  remove_mentions = TRUE,
  remove_emojis = TRUE,
  remove_punctuation = TRUE,
  remove_digits = TRUE,
  remove_url = TRUE,
  clean_spaces = TRUE,
  in_parallel = TRUE
)
}
\arguments{
\item{data}{A tibble or data frame object}

\item{text_var}{The text or message variable to analyse}

\item{tolower}{Whether to lower case the text variable or not?}

\item{remove_hashtags}{Should the user remove hashtags?}

\item{remove_mentions}{Should the user remove any user/profile mentions?}

\item{remove_emojis}{Should the user remove emojis?}

\item{remove_punctuation}{Should the user remove punctuation?}

\item{remove_digits}{Should the user remove digits?}

\item{remove_url}{Remove most forms of URL from the text variable?}

\item{clean_spaces}{Clean any redundant or pointless spaces in the text variable?}

\item{in_parallel}{Whether to run the function in parallel (TRUE = faster)}
}
\value{
The data object provided, with a cleaned text variable
}
\description{
A function that performs a series of cleaning steps on a text variable. Useful for data processing when dealing with qualitative data
}
\examples{
data <- data \%>\% data clean_text_var(text_var = message, tolower = TRUE, hashtags = TRUE, mentions = TRUE, emojis = TRUE, punctuation = TRUE, digits = TRUE, remove_url = TRUE, clean_spaces = TRUE in_parallel = TRUE)
}
