% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_group_sentiment.R
\name{count_group_sentiment}
\alias{count_group_sentiment}
\title{count_group_sentiment}
\usage{
count_group_sentiment(
  data = data,
  group_var = group_var,
  sentiment_var = sentiment_var
)
}
\arguments{
\item{data}{A tibble or data frame object}

\item{group_var}{The group variable. For example, topic or audience}

\item{sentiment_var}{The sentiment variable showing the observations ascribed sentiment based on classification}
}
\value{
A data frame with sentiment distribution counts across each of the group variables
}
\description{
count_group_sentiment
}
\examples{
sentiment_percentages <- (data = data, group_var = audience, sentiment_var = sentiment)
count_group_sentiment(data = data, group_var = audience, sentiment_var = sentiment)
}
