% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{Counter-class}
\alias{Counter}
\alias{Counter-class}
\title{Counter class}
\description{
Counter class
}
\examples{
cnt <- Counter()
show(cnt)
cnt$add()
show(cnt)
cnt$add()
cnt$add()
show(cnt)
cnt$subtract()
show(cnt)
cnt$add(5)
show(cnt)
cnt$reset()
show(cnt)
}

