% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{examiner_opts}
\alias{examiner_opts}
\title{Examiner Options}
\format{\preformatted{<environment: 0x4cb9650> 
}}
\usage{
examiner_opts
}
\description{
This environment stores the options used in the \pkg{examiner} package.
}
\keyword{datasets}

