\name{format.answerlist}
\alias{format.answerlist}
\title{Create an answerlist object}
\usage{
\method{format}{answerlist}(x, show_solutions = FALSE)
}
\arguments{
  \item{x}{An \code{answerlist} object}

  \item{show_solutions}{\code{logical}. Format in way that
  indicates the solution.}
}
\description{
Create an answerlist object
}

