% Generated by roxygen2 (4.0.0): do not edit by hand
\name{roman}
\alias{Alph}
\alias{Roman}
\alias{alph}
\alias{roman}
\title{Formatting functions}
\usage{
roman(x)

Roman(x)

alph(x)

Alph(x)
}
\arguments{
\item{x}{A vector of integers.}
}
\description{
Some functions useful for formatting integers for lists:
\describe{
\item{\code{roman}}{i, ii, iii, iv, ...}
\item{\code{Roman}}{I, II, III, IV, ...}
\item{\code{alph}}{a, b, c, d, ..., z, aa, bb, ...}
\item{\code{Alph}}{A, B, C, D, ..., Z, AA, BB, ...}
}
}
\examples{
x <- 1:10
roman(x)
Roman(x)
alph(x)
Alph(x)
}

