% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_authors_splits.R
\name{freq_authors_splits}
\alias{freq_authors_splits}
\title{freq_authors_splits}
\usage{
freq_authors_splits(
  data = data,
  text_var = text_var,
  author_var = author_var,
  url_var = url_var,
  n = n
)
}
\arguments{
\item{data}{A tibble or data frame object}

\item{text_var}{The text or message variable to examine}

\item{author_var}{The variable that contains the senders' screen name or author name}

\item{url_var}{The column with each of the posts' respective permalink or url}

\item{n}{How many of the most active users in the data set to inspect}
}
\value{
Prints the n number of most frequent authors' along with a selection of their respective message variable in the console to inspect their content, spam posters, bots or irrelevant data should be
}
\description{
Investigate the most active users within a data set, inspecting message variable for contents. A function that makes use of some dplyr functionality, grouping by author, counting and then pulling the message variable for each of the authors, printing the content in the console for investigation
}
\examples{
freq_authors_splits(data = data, text_var = message, author_var = sender_screen_name, url_var = permalink, n = 10)
freq_authors_top10 <- freq_authors_splits(data = data, text_var = message, author_var = sender_screen_name, url_var = permalink, n = 10)
}
