% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_conversations_splits.R
\name{freq_conversations_splits}
\alias{freq_conversations_splits}
\title{freq_conversations_splits}
\usage{
freq_conversations_splits(
  data = data,
  text_var = text_var,
  conversation_id = conversation_id,
  url_var = url_var,
  date = date,
  n = n
)
}
\arguments{
\item{data}{A tibble or data frame object}

\item{text_var}{The text or message variable to examine}

\item{conversation_id}{The variable that contains the conversation id}

\item{url_var}{The column with each of the posts' respective permalink or URL}

\item{date}{The date variable to display in the group splits}

\item{n}{How many of the most frequent conversation ids to inspect}
}
\value{
Prints the message variable in the console of the top n conversations in the data in terms of volume, along with the date of when the message was posted. This might be done to inspect data contents for relevance and cleanliness
}
\description{
Investigate the most common conversation ids in the data set. A function that makes use of some dplyr functionality, grouping by conversation id, counting and then pulling the message variable for each of the conversation, printing the content in the console for investigation along with the date
}
\examples{
freq_conversations_splits(data = data, text_var = message, conversation_id = conversation_id, url_var = permalink, date = date, n = 10)
freq_conversations_top10 <- freq_conversations_splits(data = data, text_var = message, conversation_id = conversation_id, url_var = permalink, date = date, n = 10)
}
