% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_column_classes.R
\name{get_column_classes}
\alias{get_column_classes}
\title{get_column_classes}
\usage{
get_column_classes(data)
}
\arguments{
\item{data}{A tibble or data frame object}
}
\value{
A data frame that displays each of the variables inside of the data object and their respective class
}
\description{
get_column_classes
}
\examples{
get_column_classes(data = data)
}
