\docType{data}
\name{opts_examiner}
\alias{opts_examiner}
\title{\code{examiner} options}
\format{\preformatted{List of 4
 $ get    :function (name, default = FALSE, drop = TRUE)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 18 9 25 3 9 3 18 25
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x311ccb8> 
 $ set    :function (...)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 26 9 33 3 9 3 26 33
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x311ccb8> 
 $ merge  :function (values)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 34 11 34 55 11 55 34 34
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x311ccb8> 
 $ restore:function (target = value)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 35 13 35 56 13 56 35 35
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x311ccb8> 
}}
\usage{
opts_examiner
}
\description{
This acts like \code{\link[knitr]{opts_knitr}} in
\pkg{knitr}.
}
\keyword{datasets}

