% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentage_distribution.R
\name{percentage_distribution}
\alias{percentage_distribution}
\title{percentage_distribution}
\usage{
percentage_distribution(data, group_var, rows_n = NULL)
}
\arguments{
\item{data}{A tibble or data frame object}

\item{group_var}{The group in which one wishes to get the \% and raw counts for, this could be sentiment, topic or brand.}

\item{rows_n}{State how many rows of data in total for if the number of rows to divide the group variable by, differs from number of rows in the \code{data} object provided.}
}
\value{
Prints the percentage distribution of any given variable
}
\description{
percentage_distribution
}
\examples{
percentage_distribution(data = data, sentiment_var = sentiment)
}
