% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dual_axis_vot.R
\name{plot_dual_axis_vot}
\alias{plot_dual_axis_vot}
\title{plot_dual_axis_vot}
\usage{
plot_dual_axis_vot(
  data = data,
  date_var = date_var,
  variable_one = variable_one,
  variable_two = variable_two,
  custom_colours = custom_colours,
  unit = c("1 year", "1 quarter", "1 month", "2 weeks", "1 week", "1 day")
)
}
\arguments{
\item{data}{A tibble or data frame object}

\item{date_var}{The date variable ascribed to each observation}

\item{variable_one}{The first variable the user wants to plot}

\item{variable_two}{The second variable, this one will be plotted on the second y axis}
}
\value{
A ggplot, with two y axes showing the trend of both variables over time
}
\description{
Plot two variables over time using two y axes
}
\examples{
data \%>\% plot_dual_axis_vot(date_var = date, variable_one = cost, variable_two = impressions)
}
