% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_group_sentiment.R
\name{plot_group_sentiment}
\alias{plot_group_sentiment}
\title{plot_group_sentiment}
\usage{
plot_group_sentiment(
  data = data,
  group_var = topic,
  sentiment_var = sentiment,
  colours = NULL
)
}
\arguments{
\item{data}{A tibble or data frame object}

\item{group_var}{The group variable. For example, topic or audience}

\item{sentiment_var}{The sentiment variable showing the observations ascribed sentiment based on classification}

\item{colours}{For if the user wishes to amend the existing colour scale}
}
\value{
A simple bar chart showing how any group variables' sentiment is distributed
}
\description{
Plots a stacked bar chart showing the sentiment distribution of a grouped variable. A volume count is also included, showing how many observations per group make up the sentiment distribution
}
\examples{
topics_sentiment_distribution <- plot_group_sentiment(data = data, group_var = topic, sentiment_var = sentiment, colours = sentiment_colours)
plot_group_sentiment(data = data, group_var = topic, sentiment_var = sentiment, colours = sentiment_colours)
}
