% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_group_vot.R
\name{plot_group_vot}
\alias{plot_group_vot}
\title{plot_group_vot}
\usage{
plot_group_vot(
  data = data,
  group_var,
  date_var,
  unit = c("day", "week", "month", "quarter", "year"),
  nrow = 2,
  plot_type = c("facet", "list"),
  colour_mapping = NULL
)
}
\arguments{
\item{data}{A tibble or data frame object}

\item{group_var}{The group variable. For example, topic or audience}

\item{date_var}{The date variable ascribed to each observation}

\item{unit}{The unit of time the user wants to display when plotting, options are; "day", "week", "month", "quarter", "year"}

\item{nrow}{The number of rows that the group variables will plotted on for when the user selects plot_type = "facet", is not required when calling "list"}

\item{plot_type}{The user has an option to select a facet plot or to plot each group variable vot individually, a list of plots. This argument is supported by "facet" or "list"}

\item{colour_mapping}{For if there is an alternative colour scale that could be used to colour the group variable}
}
\value{
A ggplot showing each group variables volume of time, either as a facetted plot or a list of plots
}
\description{
Plots a grouo variables volume over time(vot). There are two options for the user; facet or list.
}
\examples{
plot_group_vot(data = data, group_var = group, date_var = date, unit = "day", nrow = 5, plot_type = "facet", colour_mapping = colour_map)
plot_list <- plot_group_vot(data = data, group_var = group, date_var = date, unit = "day", plot_type = "list", colour_mapping = colour_map)
}
