% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_group_vot_sentiment.R
\name{plot_group_vot_sentiment}
\alias{plot_group_vot_sentiment}
\title{plot_group_vot_sentiment}
\usage{
plot_group_vot_sentiment(
  data = data,
  sentiment_var = sentiment,
  date_var = date,
  group_var = group,
  unit = c("day", "week", "month", "quarter", "year")
)
}
\arguments{
\item{data}{A tibble or data frame object}

\item{sentiment_var}{The sentiment variable showing the observations ascribed sentiment based on classification}

\item{date_var}{The date variable ascribed to each observation}

\item{group_var}{The group variable. For example, topic or audience}

\item{unit}{The unit of time the user wants to display when plotting, options are; "day", "week", "month", "quarter", "year"}
}
\value{
A faceted plot displaying volume and sentiment over time per group
}
\description{
Plots a faceted graphic, showing volume over time as well as sentiment per group
}
\examples{
topics_vot_sentiment_day <- plot_group_vot_sentiment(daat = data, sentiment_var = sentiment, date_var = date, group_var = topic, unit = "day")
plot_group_vot_sentiment(data = data sentiment_var = sentiment, date_var = date, group_var = topic, unit = "day")
}
