% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sentiment_over_time.R
\name{plot_sentiment_over_time}
\alias{plot_sentiment_over_time}
\title{plot_sentiment_over_time}
\usage{
plot_sentiment_over_time(
  data = data,
  date_var = date_var,
  sentiment_var = sentiment,
  sentiment_colors = sentiment_colors,
  vline_date = NULL,
  unit = c("day", "week", "month", "quarter", "year"),
  ...
)
}
\arguments{
\item{data}{A tibble or data frame object}

\item{date_var}{The date variable ascribed to each observation}

\item{sentiment_var}{The sentiment variable showing the observations ascribed sentiment based on classification}

\item{sentiment_colors}{For if the user has a custom colour scale or map they wish to assign each sentiment}

\item{vline_date}{Where to include a dashed line to mark a specific point on the x axis, if required}

\item{unit}{The unit of time the user wants to display when plotting, options are; "1 day", "1 week","2 weeks", "1 month", "1 year"}

\item{...}{Input argument}
}
\value{
A stacked bar chart that maps sentiment distribution per day OR week OR month OR year over time
}
\description{
plot_sentiment_over_time
}
\examples{
plot_sentiment_over_time(data = data, date_var = date, sentiment_var = sentiment, sentiment_colors = colour_map, vline_date = "2023-04-01", unit = "day")
}
