% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volume_over_time.R
\name{plot_volume_over_time}
\alias{plot_volume_over_time}
\title{plot_volume_over_time}
\usage{
plot_volume_over_time(
  data,
  date_var = date_var,
  smooth = FALSE,
  vline_date = NULL,
  bar_colour = bar_colour,
  unit = c("day", "week", "month", "quarter", "year"),
  ...
)
}
\arguments{
\item{data}{A tibble or data frame object}

\item{date_var}{The date variable ascribed to each observation}

\item{smooth}{To include a trend line utilizing geom_smooth?}

\item{vline_date}{Where to include a dashed line to mark a specific point on the x axis, if required}

\item{bar_colour}{What colour should the bar chart be? Provide a hexcode or accepted colour name}

\item{unit}{The unit of time the user wants to display when plotting, options are; "1 day", "1 week","2 weeks", "1 month", "1 year"}

\item{...}{Input argument}
}
\value{
A bar chart showing basic volume over time across the data
}
\description{
plot_volume_over_time
}
\examples{
plot_volume_over_time(data = data, date_var = date, smooth = FALSE, vline_date = "2023-04-01", bar_colour = colour_map, unit = "day")
}
