% Generated by roxygen2 (4.0.0): do not edit by hand
\name{problemblock}
\alias{format.problemblock}
\alias{problemblock}
\title{\code{problemblock} object}
\usage{
problemblock(problems, pretext = "", posttext = "", randomizable = FALSE)

\method{format}{problemblock}(x, shuffle_problems = FALSE,
  shuffle_answers = FALSE, show_solutions = FALSE,
  tpl_problemblock = examiner_opts$tpl_problemblock,
  tpl_problem = examiner_opts$tpl_problem,
  tpl_answerlist = examiner_opts$tpl_answerlist,
  format_cnt_problem_1 = identity, cnt_problem_1 = 1L,
  counter = Counter(), .debug = FALSE, ...)
}
\arguments{
\item{problems}{list of problems}

\item{pretext}{\code{character} vector with text to go before the problems.}

\item{posttext}{\code{character} vector with text to go after the problems.}

\item{randomizable}{\code{logical} Whether the problems within the block can be shuffled.}

\item{x}{The object}

\item{shuffle_problems}{Shuffle problems?}

\item{shuffle_answers}{Shuffle answers?}

\item{show_solutions}{Display the solutions to the answers?}

\item{tpl_problemblock}{Whisker template to use when rendering the object.}

\item{tpl_answerlist}{The template to use to render the object.}

\item{tpl_problem}{Whisker template to use when rendering the object.}

\item{format_cnt_problem_1}{A function used to format}

\item{cnt_problem_1}{Problem number. This will usually be set by \code{format.problem}.}

\item{counter}{Counter object used for keeping track of the total number of problems in a problemset.
Usually used internally by \code{format.problem}.}

\item{.debug}{Useful for debugging. Shows the data that is passed to the template.}

\item{...}{Used by the template and passed to \code{format.problem} for each problem.}
}
\value{
A \code{problemblock} object.
}
\description{
\code{problemblock} object
}
\details{
A \code{"problemblock"} object is a \code{"list"} with elements
\describe{
\item{\code{problems}}{The problems. A \code{"list"} of \code{"problem"} objects.}
\item{\code{pretext}}{The text to go before the problems.}
\item{\code{posttext}}{The text to go after the problems.}
}
}

