% Generated by roxygen2 (4.0.0): do not edit by hand
\name{problemset_from_list}
\alias{problemset_from_list}
\title{Create problemset from yaml file}
\usage{
problemset_from_list(x)
}
\arguments{
\item{x}{A \code{list}}
}
\value{
A \code{problemset} object.
}
\description{
Create problemset from yaml file
}

