\name{shuffle}
\alias{shuffle}
\title{Shuffle a vector}
\usage{
shuffle(x, first = 0, last = 0)
}
\arguments{
  \item{x}{vector to shuffle}

  \item{first}{Do not shuffle the first \code{first}
  elements constant.}

  \item{last}{Do not shuffle last \code{last} elements
  constant.}
}
\value{
The suffled vector.
}
\description{
Shuffle a vector, optionally holding some of the first and
last observations constant.
}

