% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_axis_app.R
\docType{data}
\name{ui}
\alias{ui}
\title{dual_axis_app}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 4.
}
\usage{
ui
}
\arguments{
\item{input}{ui}

\item{output}{server}
}
\value{
An interactive web application hosted in the users R console
}
\description{
Loads an app for the user to select a csv file from storage, select two variables as well as any colours that they wish to plot over time. The app has the same functionality as plot_dual_axis_plot, but provides the user with some control over the output
}
\examples{
runApp('R/dual_axis_app.R')
}
\keyword{datasets}
