% Generated by roxygen2 (4.0.0): do not edit by hand
\name{problem}
\alias{format.problem}
\alias{problem}
\title{Create a \code{problem} object}
\usage{
problem(text = "", answers = character(), solution = "", correct = 1L,
  first = 0L, last = 0L, randomizable = TRUE)

\method{format}{problem}(x, shuffle_answers = FALSE, show_solutions = FALSE,
  tpl_problem = examiner_opts$tpl_problem, counter = Counter(),
  cnt_problem_2 = NULL, cnt_problem_1 = 1L,
  format_cnt_problem_1 = identity, format_cnt_problem_2 = identity,
  format_cnt_problem_0 = identity, .debug = FALSE, ...)
}
\arguments{
\item{text}{\code{character} vector with the prompt for the problem.}

\item{answers}{A \code{list} of \code{problem} or \code{problemset} objects.}

\item{correct}{Indices of answers which are correct.}

\item{first}{Number of observations at the beginning of \code{answers} which will not be shuffled.}

\item{last}{Number of observations at the end of \code{answers} which will not be shuffled.}

\item{solution}{Text for the solution to the problem.}

\item{randomizable}{Can the answers be shuffled?}

\item{x}{The object}

\item{shuffle_answers}{Shuffle answers?}

\item{show_solutions}{Display the solutions to the answers?}

\item{tpl_problem}{Whisker template to use when rendering the object.}

\item{counter}{A counter to keep track of the number of problems.}

\item{cnt_problem_1}{A counter for the top level \code{problem} and \code{problemset} objects.}

\item{cnt_problem_2}{\code{NULL} if the problem is not in a \code{problemblock}. Otherwise, the number within that \code{problemblock}.}

\item{format_cnt_problem_0}{A function used to format \code{cnt_problem_0}.}

\item{format_cnt_problem_1}{A function used to format \code{cnt_problem_1}.}

\item{format_cnt_problem_2}{A function used to format \code{cnt_problem_2}.}

\item{.debug}{Useful for debugging. Shows the data that is passed to the template.}

\item{...}{Used by the template and passed to \code{format.answerlist}.}
}
\value{
An object of class \code{problem}.
}
\description{
Create a \code{problem} object
}
\details{
The class \code{"problem"} is a \code{"list"} with elements,
\describe{
\item{\code{text}}{\code{"character"}. The text for the problem.}
\item{\code{answers}}{\code{"answerlist"}. Possible answers to the problem.}
\item{\code{solution}}{\code{"character"}. Text with discussion of the answer to problem.}
\item{\code{randomizable}}{\code{"logical"}. Whether the answers can be shuffled.}
}
}

