\name{problemblock}
\alias{problemblock}
\title{Create a problemblock object}
\usage{
problemblock(problems, pretext = "", posttext = "", randomizable = FALSE)
}
\arguments{
  \item{problems}{list of problems}

  \item{pretext}{\code{character} vector with text to go
  before the problems.}

  \item{posttext}{\code{character} vector with text to go
  after the problems.}

  \item{randomizable}{\code{logical} Whether the problems
  within the block can be shuffled.}
}
\value{
A \code{problemblock} object.
}
\description{
Create a problemblock object
}

