\name{problemset}
\alias{problemset}
\title{Create a problemset object}
\usage{
problemset(x, problems)
}
\arguments{
  \item{x}{\code{character} vector with the prompt for the
  problemset.}

  \item{problems}{list of problems or other problemsets}
}
\value{
A \code{problemset} object
}
\description{
Create a problemset object
}

