% Generated by roxygen2 (4.0.0): do not edit by hand
\name{problemset}
\alias{problemset}
\title{Create a problemset object}
\usage{
problemset(problems, pretext = "", posttext = "")
}
\arguments{
\item{problems}{list of problems}

\item{pretext}{\code{character} vector with text to go before the problems.}

\item{posttext}{\code{character} vector with text to go after the problems.}
}
\value{
A \code{problemset} object.
}
\description{
Create a problemset object
}

