% Generated by roxygen2 (4.0.0): do not edit by hand
\name{problemset}
\alias{format.problemset}
\alias{problemset}
\title{\code{problemset} object}
\usage{
problemset(problems, pretext = "", posttext = "")

\method{format}{problemset}(x, shuffle_problems = FALSE,
  shuffle_answers = FALSE, show_solutions = FALSE,
  tpl_problemset = examiner_opts$tpl_problemset, .debug = FALSE, ...)
}
\arguments{
\item{problems}{list of problems}

\item{pretext}{\code{character} vector with text to go before the problems.}

\item{posttext}{\code{character} vector with text to go after the problems.}

\item{x}{The object.}

\item{shuffle_problems}{Shuffle problems?}

\item{shuffle_answers}{Shuffle answers?}

\item{show_solutions}{Display the solutions to the answers?}

\item{tpl_problemset}{Whisker template to use when rendering the object.}

\item{.debug}{Useful for debugging. Shows the data that is passed to the template.}

\item{...}{Used by the template and passed to \code{format.problem} and \code{format.problemblock}.}
}
\value{
A \code{problemset} object.
}
\description{
\code{problemset} object
}
\details{
A \code{"problemset"} object is a \code{"list"} with elements
\describe{
\item{\code{problems}}{The problems. A \code{"list"} of \code{"problem"} objects.}
\item{\code{pretext}}{The text to go before the problems.}
\item{\code{posttext}}{The text to go after the problems.}
}
}

